Do not use clusters_handler for handling management actions
