+++ handlers/admin/clusters/clusters_handler.py
@@ -29,7 +29,6 @@ class AdminClusterStates(StatesGroup):
     waiting_for_inbound_id = State()
     waiting_for_server_name = State()
     waiting_for_subscription_url = State()
-    waiting_for_new_domain = State()
 
 
 @router.callback_query(
@@ -370,52 +369,3 @@ async def handle_clusters_sync(
         await callback_query.message.answer(
             text=f"❌ Произошла ошибка при синхронизации: {e}", reply_markup=build_admin_back_kb("clusters")
         )
-
-
-@router.callback_query(AdminClusterCallback.filter(F.action == "change_domain"), IsAdminFilter())
-async def request_new_domain(callback_query: AdminClusterCallback, state: FSMContext):
-    """Запрашивает у администратора новый домен."""
-    await state.set_state(AdminClusterStates.waiting_for_new_domain)
-    await callback_query.message.edit_text(
-        text="🌐 Введите новый домен (без https://):\nПример: pocomachodomen.ru",
-    )
-
-
-@router.message(AdminClusterStates.waiting_for_new_domain)
-async def process_new_domain(message: Message, state: FSMContext, session: asyncpg.Connection):
-    """Обновляет домен в таблице keys."""
-    new_domain = message.text.strip()
-    logger.info(f"[DomainChange] Новый домен, введённый администратором: '{new_domain}'")
-
-    if not new_domain or " " in new_domain or not new_domain.replace(".", "").isalnum():
-        logger.warning("[DomainChange] Некорректный домен")
-        await message.answer(
-            "🚫 Некорректный домен! Введите домен без http:// и без пробелов.",
-            reply_markup=build_admin_back_kb("admin"),
-        )
-        return
-
-    new_domain_url = f"https://{new_domain}"
-    logger.info(f"[DomainChange] Новый домен с протоколом: '{new_domain_url}'")
-
-    query = """
-        UPDATE keys
-        SET key = regexp_replace(key, '^https://[^/]+', $1::TEXT)
-        WHERE key NOT LIKE $1 || '%'
-    """
-    try:
-        await session.execute(query, new_domain_url)
-        logger.info("[DomainChange] Запрос на обновление домена выполнен успешно.")
-    except Exception as e:
-        logger.error(f"[DomainChange] Ошибка при выполнении запроса: {e}")
-        await message.answer(f"❌ Ошибка при обновлении домена: {e}", reply_markup=build_admin_back_kb("admin"))
-        return
-
-    try:
-        sample = await session.fetchrow("SELECT key FROM keys LIMIT 1")
-        logger.info(f"[DomainChange] Пример обновленной записи: {sample}")
-    except Exception as e:
-        logger.error(f"[DomainChange] Ошибка при выборке обновленной записи: {e}")
-
-    await message.answer(f"✅ Домен успешно изменен на {new_domain}!", reply_markup=build_admin_back_kb("admin"))
-    await state.clear()
