+++ handlers/admin/management/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ("router",)
+
+from management_handler import router
